src/requirements.txt
greenlet==2.0.2
importlib-metadata==6.0.0
importlib-resources==5.12.0
Mako==1.2.4
MarkupSafe==2.1.2
mysql-connector==2.2.9
python-dotenv==0.21.1
SQLAlchemy==2.0.4
typing_extensions==4.5.0
zipp==3.14.0
